pyproject.toml
[project]
name = "substack-stacker"
version = "0.1.0"
description = "Compile a Substack publication into a single .docx book"
requires-python = ">=3.9"
dependencies = [
    "python-docx>=1.1.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "tqdm>=4.66.0",
]

[project.scripts]
substack-stacker = "substack_stacker.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
